uploading the webserver cgi loop class update
