+++ src/main.hpp
@@ -30,6 +30,7 @@
 # include "response.hpp"
 # include "serverFd.hpp"
 # include "webserver.hpp"
+# include "loopHandler.hpp"
 # include <algorithm>    // std::count
 
 std::string get_actual_date();
